gfx906: cherry-pick 21a0e85f feat(kv-cache): fp16 V storage and PV compute (upstream)
Upstream stores the V half of every KV cache as FP16_RNE(bf16 input) and runs
PV in fp16; K stays bf16. The port's tree predates upstream's softmax_attention/
kv_cache split, so the pick was applied as a patch and re-expressed on the
port's own files.

Taken from upstream as-is: cyclic_kv_cache.{h,cpp} (V layer dtype FP16),
launcher/wrapper casts and dtype checks for swa + bidirectional_gqa_attention.

Port adaptations (gfx906 kernels read V from the cache themselves):
- ops/common/math.cuh: pack_f16x2 / bf16x2_bits_to_f16x2_bits /
  bf16x8_bits_to_f16x8_bits added after the compat block; bf16->f32 is an exact
  shift, f32->f16 uses __floats2half2_rn (no bf16 VALU needed).
- kernel/gqa_attention_decode_bf16_gfx906.cuh (wave64 SIMT small-T decode):
  cache_v is __half*; the append phase converts new-token V with
  bf16x8_bits_to_f16x8_bits; resolve_v always reads the cache (the append
  phase covers the split range before the __syncthreads, mirroring upstream's
  tc kernel); PV loads __half2 and accumulates in fp32 as before.
- kernel/gqa_attention_prefill_bf16_gfx906.cuh (wave64 SIMT prefill): cache_v
  const __half*, v_s smem __half, __half2 loads, fp32 accumulate.
- kernel/gqa_attention_prefill_bf16.cuh: gqa_attention_prefill_fill_bf16_kernel
  (shared fill kernel) converts V on store; launcher casts updated.
- kernel/kv_cache_append_prefix.cuh + launcher/wrapper: cyclic and paged units
  store FP16_RNE(V); wrappers require V plane dtype FP16 (K stays BF16).
- wrapper/gqa_attention.cpp: bf16 cache requires k_pages BF16 / v_pages FP16
  (int8 cache unchanged: both planes I8).
- state/decoder_state.cpp + runtime/layouts_impl.h (dflash full pool): V plane
  planned as FP16 for bf16 caches; int8-G64 planes unchanged.
- core/hip_compat.h: cudaStreamCreate, cudaStreamCaptureModeGlobal, and the
  five-argument cudaGraphInstantiate form (test tooling) mapped to HIP.
- kernel/bidirectional_gqa_attention.cuh (dflash tensor-core kernel, not used on
  gfx906): context_v is const __half* and is staged bit-for-bit; upstream's
  fp16 query-V conversion + mma_f16 PV is NOT mirrored here (noted in-source).
  tests/ops/test_swa.cpp and test_bidirectional_gqa_attention.cpp kept at HEAD
  (not built on the port).
- Upstream's fp32 partial_acc change for the split-KV reducer was NOT taken:
  the port's reducer/partials stay bf16 (scope: consume fp16 V only).
- tests/ops/test_gqa_attention.cpp and test_kv_cache_append_prefix.cpp mirror
  upstream's reference changes (f16 history/oracle, finite bf16 input bits).
- Upstream-only files (softmax_attention/*, kv_cache/append/*, paged_kv_storage.h,
  new public headers, docs, benches) dropped.

Gates on card 2 (bf16 KV): ninfer_gqa_attention_test OK,
ninfer_kv_cache_append_prefix_test OK; p4 MTP d3 128-new and p3 plain 64-new
greedy stdout byte-identical to 52e3c864; deep-ctx probe v3 30K through
ninfer-serve PASS 7/7; quick bench pp512/pp2048/tg128 within noise of post-1d.

+++ src/ops/launcher/gqa_attention_decode.cu
@@ -110,7 +110,7 @@ void launch_tc_partial_bf16(const Tensor& q, CacheInput input, const Tensor& pos
 #endif
         static_cast<const __nv_bfloat16*>(q.data), input,
         static_cast<const std::int32_t*>(pos.data), static_cast<__nv_bfloat16*>(cache_k.data),
-        static_cast<__nv_bfloat16*>(cache_v.data),
+        static_cast<__half*>(cache_v.data),
         static_cast<const std::int32_t*>(cache.block_tables.data),
         invocation.valid_columns == nullptr
             ? nullptr
