+++ src/ops/launcher/kv_cache_append_prefix.cu
@@ -24,7 +24,7 @@ void launch_paged(const Tensor& k, const Tensor& v, const Tensor& positions, con
     validate_plan(k, plan);
     if (plan.max_count == 0) return;
     auto* cache_k       = static_cast<__nv_bfloat16*>(cache.k_pages.data);
-    auto* cache_v       = static_cast<__nv_bfloat16*>(cache.v_pages.data);
+    auto* cache_v       = static_cast<__half*>(cache.v_pages.data);
     const auto* input_k = static_cast<const __nv_bfloat16*>(k.data);
     const auto* input_v = static_cast<const __nv_bfloat16*>(v.data);
     const auto* pos     = static_cast<const std::int32_t*>(positions.data);
@@ -45,7 +45,7 @@ void launch_cyclic(const Tensor& k, const Tensor& v, const Tensor& positions, co
     validate_plan(k, plan);
     if (plan.max_count == 0) return;
     auto* cache_k       = static_cast<__nv_bfloat16*>(cache.k.data);
-    auto* cache_v       = static_cast<__nv_bfloat16*>(cache.v.data);
+    auto* cache_v       = static_cast<__half*>(cache.v.data);
     const auto* input_k = static_cast<const __nv_bfloat16*>(k.data);
     const auto* input_v = static_cast<const __nv_bfloat16*>(v.data);
     const auto* pos     = static_cast<const std::int32_t*>(positions.data);
